Move dev only JS hooks to separate file
